refactor: split ui package
